refactor larkservice package path
