+++ pkg/microservice/aslan/core/system/service/lark/lark.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package service
+package lark
 
 import (
 	"context"
